Gradle 8.12 requires newer versions of Windows (not mentioned in release notes). gradle/gradle#31939
For this upgrade we may just want to get on Gradle 8.11. Upgrading from Windows 2016 to 2019 seems good as 2016 is out of mainline support by Microsoft, but we'd do that separately. I see there is already a Windows 2019 pool for us to use, but we'll need to see how involved the upgrade is.