Revert "xds: Convert CdsLb to XdsDepManager"
This reverts commit 297ab05efeb0565c195518a819dfa851d1c0d62b.

b/430347751 shows multiple concerning behaviors in the xDS stack with
the new A74 config update model. XdsDepManager and CdsLB2 still seem to
be working correctly, but the change is exacerbated issues in other
parts of the stack, like RingHashConfig not having equals fixed in
a8de9f07ab.

Revert only for the v1.74.x release, leaving it on master.