SECURITY.md: Mention gcompat for Alpine
In the olden days Alpine didn't work at all. Then it worked. And then
sometime in 2021 it started failing (#8751) because of missing __strndup.
The most recent deep investigation showed it is missing __strdup these
days https://github.com/grpc/grpc-java/issues/11660#issuecomment-2469284111 .

I'm not expecting too many people to find this themselves, but we can
link to it when they experience problems.